Initialize devtools automatically
Initialize the React Developer Tools integration automatically when
'preact/devtools' is required.

This makes for simpler usage for most users. Anyone who needs to be able
to unregister the devtools integration, such as the tests, can reach
into the module and invoke `initDevTools()` directly.
